Create ivo_goes_to_citycenter.rb
Created little Control Flow exercize based a live situation of a colleague who was asking if someone would give him a lift after work. 

Apologies if I am commiting this to the master branch. I didn-t have the time to study the "fork" handling and am very enthusiastic to share this with you. xxx